Split loop functions for better control and protection on the read functions
