+++ incs/minishell.h
@@ -1,6 +1,8 @@
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
+# define MAX_INPUT_SIZE 131072
+
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <unistd.h>
@@ -10,6 +12,9 @@
 #include "libft.h"
 
 void    ft_print(char *s);
+void    ft_putstr(char *s);
 void    print_prompt_prefix(void);
+void	prompt(void);
+void	launch(char *input);
 
 #endif
